ЄMove youtube screen to a separated module
