+++ app/build.gradle
@@ -85,6 +85,7 @@ dependencies {
     implementation "com.google.dagger:hilt-android:$hilt_version"
     implementation project(path: ':feature:download-manager')
     implementation project(path: ':feature:saved-music')
+    implementation project(path: ':feature:youtube-downloader')
     kapt "com.google.dagger:hilt-compiler:$hilt_version"
     implementation "androidx.work:work-runtime-ktx:$work_version"
     implementation "androidx.hilt:hilt-work:$androidx_hilt_work_version"
