--- app/src/main/java/com/yurii/youtubemusic/screens/youtube/Mapping.kt
@@ -1,7 +1,0 @@
-package com.yurii.youtubemusic.screens.youtube
-
-import com.google.api.services.youtube.model.Video
-import com.youtubemusic.core.model.VideoItem
-
-import org.threeten.bp.Duration
-
