+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".screens.main.MainActivity">
+    tools:context=".MainActivity">
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:id="@+id/coordinator_layout"
@@ -57,7 +57,7 @@
             app:itemIconTint="@color/nav_item_state"
             app:itemTextColor="@color/nav_item_state"
             app:labelVisibilityMode="selected"
-            app:layout_behavior="com.yurii.youtubemusic.ui.BottomNavigationBehavior"
+            app:layout_behavior="com.yurii.youtubemusic.utils.BottomNavigationBehavior"
             app:menu="@menu/navigation_menu" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
