--- app/src/main/res/layout/fragment_youtube_music.xml
@@ -1,158 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".screens.youtube.YouTubeMusicFragment">
-
-    <data>
-
-        <variable
-            name="viewState"
-            type="com.yurii.youtubemusic.screens.youtube.YouTubeMusicFragment.ViewState" />
-
-        <import type="com.yurii.youtubemusic.screens.youtube.YouTubeMusicFragment.ViewState" />
-    </data>
-
-    <com.youtubemusic.core.common.ui.NestedScrollCoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/white"
-        android:clipToPadding="false">
-
-
-        <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/layout_selection_playlist"
-            isVisible="@{!(viewState instanceof ViewState.NoSelectedPlaylist)}"
-            android:layout_width="match_parent"
-            android:background="?attr/colorPrimary"
-            android:layout_height="wrap_content">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:paddingLeft="@dimen/mediumPadding"
-                android:paddingRight="@dimen/mediumPadding"
-                app:layout_scrollFlags="scroll|enterAlways|snap"
-                app:popupTheme="@style/ThemeOverlay.AppCompat.Light">
-
-                <TextView
-                    android:id="@+id/tvPlayListName"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="1"
-                    android:ellipsize="end"
-                    android:maxLines="1"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Title"
-                    android:textColor="@android:color/white"
-                    android:textSize="12sp"
-                    tools:text="@tools:sample/lorem/random" />
-
-                <com.google.android.material.button.MaterialButton
-                    android:id="@+id/btn_select_play_list"
-                    style="@style/Widget.AppCompat.Button.Borderless"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/mediumMargin"
-                    android:minHeight="0dp"
-                    android:text="@string/select" />
-            </LinearLayout>
-
-        </com.google.android.material.appbar.AppBarLayout>
-
-        <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-            android:id="@+id/refresh"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:gravity="center"
-                android:orientation="vertical">
-
-
-                <androidx.recyclerview.widget.RecyclerView
-                    android:id="@+id/videos"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fitsSystemWindows="true"
-                    isVisible="@{viewState instanceof ViewState.VideosLoaded}"
-                    tools:listitem="@layout/item_video"
-                    tools:visibility="visible" />
-
-
-                <TextView
-                    android:id="@+id/label_empty_playlist"
-                    style="@style/DescriptionText"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/label_empty_playlist"
-                    android:textAlignment="center"
-                    isVisible="@{viewState instanceof ViewState.EmptyList}"
-                    app:drawableTopCompat="@drawable/ic_empty_playlist"
-                    tools:visibility="gone" />
-
-
-                <ProgressBar
-                    android:id="@+id/loading"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    isVisible="@{viewState instanceof ViewState.Loading}"
-                    tools:visibility="gone" />
-
-                <LinearLayout
-                    android:id="@+id/layout_selection_first_playlist"
-                    isVisible="@{viewState instanceof ViewState.NoSelectedPlaylist}"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:gravity="center"
-                    android:orientation="vertical"
-                    tools:visibility="gone">
-
-                    <TextView
-                        style="@style/DescriptionText"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="@string/label_first_playlist"
-                        app:drawableTopCompat="@drawable/ic_first_selection_illustration" />
-
-                    <com.google.android.material.button.MaterialButton
-                        android:id="@+id/btn_select_play_list_first"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginTop="@dimen/mediumMargin"
-                        android:backgroundTint="@color/colorAccent"
-                        android:text="@string/select_playlist"
-                        android:textColor="@android:color/white" />
-
-                </LinearLayout>
-
-                <LinearLayout
-                    android:id="@+id/error_layout"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:gravity="center"
-                    android:orientation="vertical"
-                    isVisible="@{viewState instanceof ViewState.Error}">
-
-                    <TextView
-                        android:id="@+id/error"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_margin="8dp"
-                        android:textAlignment="center"
-                        app:drawableTopCompat="@drawable/ic_a_broken_robot_pana" />
-
-                    <Button
-                        android:id="@+id/btn_try_again"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_margin="8dp"
-                        android:text="@string/label_try_again" />
-                </LinearLayout>
-            </LinearLayout>
-        </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-    </com.youtubemusic.core.common.ui.NestedScrollCoordinatorLayout>
-</layout>
