+++ feature/youtube-downloader/src/main/res/drawable/ic_first_selection_illustration.xml
