+++ feature/youtube-downloader/src/main/res/drawable/ic_login_illustration.xml
