+++ feature/youtube-downloader/src/main/res/layout/layout_no_categories.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/no_categories"
+    style="@style/DescriptionText"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_gravity="center_horizontal"
+    android:text="@string/label_no_playlist"
+    android:visibility="visible"
+    app:drawableTopCompat="@drawable/ic_empty_categories" />
+
