+++ settings.gradle
@@ -27,3 +27,4 @@ include ':feature:playlist-editor'
 include ':feature:download-manager'
 include ':feature:player'
 include ':feature:saved-music'
+include ':feature:youtube-downloader'
